[Mengxing Yuan] #369 feat: add requisition number

+++ src/main/java/org/siglus/siglusapi/service/SiglusRequisitionExtensionService.java
@@ -56,6 +56,9 @@ public class SiglusRequisitionExtensionService {
   public void deleteRequisitionExtension(UUID requisitionId) {
     RequisitionExtension requisitionExtension = requisitionExtensionRepository
         .findByRequisitionId(requisitionId);
+    if (null == requisitionExtension) {
+      return;
+    }
     log.info("delete requisition extension by requisition id: {}", requisitionId);
     requisitionExtensionRepository.delete(requisitionExtension);
   }
